starter/requirements.txt
numpy
pandas
scikit-learn
pytest
requests
fastapi==0.112.2
uvicorn
gunicorn
dvc==3.54.1
dvc-gs
jupyter
ydata_profiling
flake8
black
google-auth